docs(multi-action-button): add playground example
