+++ skills/carbon-react/components/multi-action-button.md
@@ -315,29 +315,6 @@ description: Carbon MultiActionButton component props and usage examples.
 | aria-grabbed | Booleanish \| undefined | No |  | Yes | in ARIA 1.1 | Indicates an element's "grabbed" state in a drag-and-drop operation. |  |
 
 ## Examples
-### DefaultStory
-
-**Args**
-
-```tsx
-{ text: "Multi Action Button" }
-```
-
-**Render**
-
-```tsx
-(args: MultiActionButtonProps) => {
-    return (
-      <MultiActionButton {...args}>
-        <Button href="#">Button 1</Button>
-        <Button>Button 2</Button>
-        <Button>Button 3</Button>
-      </MultiActionButton>
-    );
-  }
-```
-
-
 ### Focusing Main Button Programmatically
 
 **Render**
@@ -364,188 +341,6 @@ description: Carbon MultiActionButton component props and usage examples.
     </Box>
   );
 }
-```
-
-
-### Disabled
-
-**Args**
-
-```tsx
-{ ...DefaultStory.args, text: "Multi Action Button", disabled: true }
-```
-
-
-### Sizes
-
-**Render**
-
-```tsx
-() => {
-  return (["small", "medium", "large"] as const).map(
-    (size: MultiActionButtonProps["size"]) => (
-      <Box key={size} mb={3}>
-        <MultiActionButton size={size} text={`Multi Action Button - ${size}`}>
-          <Button size={size} href="#">
-            Button 1
-          </Button>
-          <Button size={size}>Button 2</Button>
-          <Button size={size}>Button 3</Button>
-        </MultiActionButton>
-      </Box>
-    ),
-  );
-}
-```
-
-
-### Custom Width
-
-**Args**
-
-```tsx
-{
-  text: "Multi Action Button",
-  width: 0.7,
-}
-```
-
-**Render**
-
-```tsx
-(args: MultiActionButtonProps) => {
-  return (
-    <MultiActionButton {...args}>
-      <Button href="#">Button 1</Button>
-      <Button>Button 2</Button>
-      <Button>Button 3</Button>
-    </MultiActionButton>
-  );
-}
-```
-
-
-### Button Types
-
-**Render**
-
-```tsx
-() => {
-  return (["primary", "secondary", "tertiary"] as const).map(
-    (buttonType: MultiActionButtonProps["buttonType"]) => (
-      <Box key={buttonType} mb={3}>
-        <MultiActionButton
-          buttonType={buttonType}
-          text={`Multi Action Button - ${buttonType}`}
-        >
-          <Button href="#">Button 1</Button>
-          <Button>Button 2</Button>
-          <Button>Button 3</Button>
-        </MultiActionButton>
-      </Box>
-    ),
-  );
-}
-```
-
-
-### Child Button Types
-
-**Render**
-
-```tsx
-() => {
-  return (
-    <MultiActionButton text="Multi Action Button">
-      <Button>Default button</Button>
-      <Button buttonType="primary">Primary</Button>
-      <Button buttonType="primary" destructive>
-        Primary - destructive
-      </Button>
-      <Button buttonType="secondary">Secondary</Button>
-      <Button buttonType="secondary" destructive>
-        Secondary - destructive
-      </Button>
-      <Button buttonType="tertiary">Tertiary</Button>
-      <Button buttonType="tertiary" destructive>
-        Tertiary - destructive
-      </Button>
-      <Button disabled>Disabled</Button>
-    </MultiActionButton>
-  );
-}
-```
-
-
-### Alignment
-
-**Render**
-
-```tsx
-() => {
-  return (["left", "right"] as const).map(
-    (align: MultiActionButtonProps["align"]) => (
-      <Box key={align} mb={3}>
-        <MultiActionButton
-          align={align}
-          text={`Multi Action Button - ${align}`}
-        >
-          <Button href="#">Button 1</Button>
-          <Button>Button 2</Button>
-          <Button>Button 3</Button>
-        </MultiActionButton>
-      </Box>
-    ),
-  );
-}
-```
-
-
-### Position
-
-**Render**
-
-```tsx
-() => {
-  return (
-    <Box display="flex" justifyContent="space-around">
-      <MultiActionButton position="left" text="Left position">
-        <Button href="#">Button 1 with longer text</Button>
-        <Button>Button 2</Button>
-        <Button>Button 3</Button>
-      </MultiActionButton>
-
-      <MultiActionButton position="right" text="Right position">
-        <Button href="#">Button 1 with longer text</Button>
-        <Button>Button 2</Button>
-        <Button>Button 3</Button>
-      </MultiActionButton>
-    </Box>
-  );
-}
-```
-
-
-### Subtext
-
-**Args**
-
-```tsx
-{
-    ...DefaultStory.args,
-    size: "large",
-    text: "Multi Action Button",
-    subtext: "subtext",
-    children: (
-      <>
-        <Button size="large" href="#">
-          Button 1
-        </Button>
-        <Button size="large">Button 2</Button>
-        <Button size="large">Button 3</Button>
-      </>
-    ),
-  }
 ```
