+++ src/components/multi-action-button/multi-action-button.mdx
@@ -1,4 +1,4 @@
-import { Meta, ArgTypes, Canvas } from "@storybook/addon-docs/blocks";
+import { Meta, ArgTypes, Canvas, Controls } from "@storybook/addon-docs/blocks";
 
 import * as ButtonStories from "../button/button.stories.tsx";
 import * as MultiActionButtonStories from "./multi-action-button.stories.tsx";
@@ -29,65 +29,36 @@ import * as MultiActionButtonStories from "./multi-action-button.stories.tsx";
 import MultiActionButton, { type MultiActionButtonHandle } from "carbon-react/lib/components/multi-action-button";
 ```
 
+## Playground
+
+Use this interactive example to explore MultiActionButton props with Storybook controls.
+
+<Canvas of={MultiActionButtonStories.Playground} />
+
+<Controls
+  of={MultiActionButtonStories.Playground}
+  include={[
+    "text",
+    "disabled",
+    "buttonType",
+    "size",
+    "subtext",
+    "width",
+    "align",
+    "position",
+  ]}
+/>
+
 ## Examples
-
-### Default
-
-<Canvas of={MultiActionButtonStories.DefaultStory} />
 
 ### Focusing Main Button Programmatically
 
-The `MultiActionButtonHandle` type provides an imperative handle for programmatic control over `MultiActionButton`. 
+The `MultiActionButtonHandle` type provides an imperative handle for programmatic control over `MultiActionButton`.
 Using a `ref`, you can access its `focusMainButton()` method to set focus on the main button as needed.
 
 <Canvas of={MultiActionButtonStories.ProgrammaticFocus} />
 
-### Disabled
-
-<Canvas of={MultiActionButtonStories.Disabled} />
-
-### Custom width
-
-Numbers from 0-1 are converted to percentage widths. Numbers greater than 1 are converted to pixel values.
-String values are passed as raw CSS values. `size` prop can still be used to set desired height.
-
-<Canvas of={MultiActionButtonStories.CustomWidth} />
-
-### Button types
-
-<Canvas of={MultiActionButtonStories.ButtonTypes} />
-
-### Child Button types
-
-Buttons used as a content of `MultiActionButton` can be of any type.
-
-<Canvas of={MultiActionButtonStories.ChildButtonTypes} />
-
-### Sizes
-
-<Canvas of={MultiActionButtonStories.Sizes} />
-
-### Button text alignment
-
-Use the `align` prop to change the alignment of the text in the menu Buttons.
-
-<Canvas of={MultiActionButtonStories.Alignment} />
-
-### Menu position
-
-By default, the menu will open aligned to the "left" of the main button, but you can use the `position` prop to change this to the "right". 
-
-Note: The position of the menu will also be automatically adjusted to fit within the viewport.
-
-<Canvas of={MultiActionButtonStories.Position} />
-
-### Subtext
-
-Subtext only works when `size` is `large`
-
-<Canvas of={MultiActionButtonStories.Subtext} />
-
-## When children buttons render icons
+### Child buttons with icons
 
 <Canvas of={MultiActionButtonStories.WithChildrenButtonsWithIcons} />
 
@@ -124,6 +95,6 @@ Subtext only works when `size` is `large`
 
 `MultiActionButton`'s forwarded ref exposes the following imperative methods:
 
-| Method Name           | Description                               |
-| --------------------- | ----------------------------------------- |
+| Method Name         | Description                               |
+| ------------------- | ----------------------------------------- |
 | `focusMainButton()` | Programmatically focuses the main button. |
